refactor(Router, constants): Path 상수 추가 #4
- Path에 해당하는 상수를 추가합니다.
- 모듈명 단수에서 복수로 변경
- 모듈 import 파일 부분 갱신
