+++ src/api/client.ts
@@ -1,5 +1,5 @@
 import axios, { AxiosRequestConfig } from "axios";
-import { baseURL } from "../constants/constant";
+import { baseURL } from "../constants/constants";
 
 const axiosConfig: AxiosRequestConfig = {
   baseURL,
