--- src/constants/constant.ts
@@ -1,3 +1,0 @@
-const baseURL = "https://www.pre-onboarding-selection-task.shop";
-
-export { baseURL };
