+++ src/constants/constants.ts
@@ -1,0 +1,14 @@
+/** 서버 origin */
+const baseURL = "https://www.pre-onboarding-selection-task.shop";
+
+const SIGNIN_PATH = "/signin";
+
+const SIGNUP_PATH = "/signup";
+
+const MAIN_PATH = "/";
+
+const TODO_PATH = "/todo";
+
+const AUTH_PATH = "/auth";
+
+export { baseURL, MAIN_PATH, SIGNUP_PATH, SIGNIN_PATH, TODO_PATH, AUTH_PATH };
